[feat_action][taier-ui]support to excute sql quickly
